fix(backend): fixed some bugs from integration

+++ dino-backend/src/main/java/com/team7/dino/controller/OrderController.java
@@ -77,7 +77,7 @@ public class OrderController {
                 .builder()
                 .id(orderId)
                 // TODO: issue here; purchasedAt being stored as negative number?
-                .purchasedAt(json.get("purchasedAt").asInt())
+                .purchasedAt(json.get("purchasedAt").asLong())
                 .total(json.get("total").asDouble())
                 .state("New")
                 .userId(removeQuotes.apply(json, "userId"))
